pyproject.toml
[project]
name = "npo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "snowflake-snowpark-python>=1.45.0",
    "pytest>=9.0.2",
]
